+++ internal/cmd/base/servers.go
@@ -596,6 +596,37 @@ func (b *Server) DestroyDevDatabase() error {
 	return nil
 }
 
+func (b *Server) SetupControllerPublicClusterAddress(conf *config.Config, flagValue string) error {
+	if conf.Controller == nil {
+		conf.Controller = new(config.Controller)
+	}
+	if flagValue != "" {
+		conf.Controller.PublicClusterAddr = flagValue
+	}
+	if conf.Controller.PublicClusterAddr == "" {
+	FindAddr:
+		for _, listener := range conf.Listeners {
+			for _, purpose := range listener.Purpose {
+				if purpose == "cluster" {
+					conf.Controller.PublicClusterAddr = listener.Address
+					break FindAddr
+				}
+			}
+		}
+	}
+	host, port, err := net.SplitHostPort(conf.Controller.PublicClusterAddr)
+	if err != nil {
+		if strings.Contains(err.Error(), "missing port") {
+			port = "9201"
+			host = conf.Controller.PublicClusterAddr
+		} else {
+			return fmt.Errorf("Error splitting public cluster adddress host/port: %w", err)
+		}
+	}
+	conf.Controller.PublicClusterAddr = net.JoinHostPort(host, port)
+	return nil
+}
+
 func (b *Server) SetupWorkerPublicAddress(conf *config.Config, flagValue string) error {
 	if conf.Worker == nil {
 		conf.Worker = new(config.Worker)