+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ## vNext
 
+### New and Improved
+
+* docker: Official Docker image for `hashicorp/boundary`
+  ([PR](https://github.com/hashicorp/boundary/pull/755))
+* controller: Add ability to set public address for cluster purposes
+  ([Issue](https://github.com/hashicorp/boundary/pull/758))
+  ([PR](https://github.com/hashicorp/boundary/pull/761))
+
 ### Bug Fixes
 
 * controller, worker: Fix IPv4-only check so `0.0.0.0` specified without a port